+++ src/caterpillar/__init__.py
@@ -14,8 +14,8 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import warnings
 
-__version__ = "2.6.3"
-__release__ = "2.6"
+__version__ = "2.7.0"
+__release__ = "2.7"
 __author__ = "MatrixEditor"